Move iconDataList to basics
